+++ voip_sip_webrtc_twilio/__manifest__.py
@@ -1,13 +1,15 @@
 # -*- coding: utf-8 -*-
 {
     'name': "Voip Communication - Twilio",
-    'version': "1.0.0",
+    'version': "1.0.1",
     'author': "Sythil Tech",
     'category': "Tools",
     'support': "steven@sythiltech.com.au",
     'summary': "Add support for Twilio XML",
     'license':'LGPL-3',
     'data': [
+        'views/voip_twilio_views.xml',
+        'views/voip_call_views.xml',
     ],
     'demo': [],
     'depends': ['voip_sip_webrtc'],