+++ voip_sip_webrtc_twilio/models/voip_call.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+import logging
+_logger = logging.getLogger(__name__)
+from openerp import api, fields, models
+
+class VoipCallInheritTWilio(models.Model):
+
+    _inherit = "voip.call"
+
+    twilio_sid = fields.Char(string="Twilio SID")