+++ voip_sip_webrtc_twilio/doc/changelog.rst
@@ -1,3 +1,7 @@
+v1.0.1
+======
+* Import call history from Twilio
+
 v1.0.0
 ======
 * Initial Release